fix: redirect cdn.../fonts/* to apex, recover stale-cache 404s (#3248)
Root cause (already fixed in app code by #1448): a ~37h window where
vite.config.ts rebased font URLs onto the CDN domain. Clients that
cached HTML from that window still hold an absolute cdn.../fonts/...
URL, which 404s forever since fonts were never intentionally hosted
on the CDN. No live repro was ever captured across 3 investigation
rounds, so this adds a Cloudflare Redirect Rule that 301s any
cdn.../fonts/* request to the apex, closing the gap at the edge
regardless of which client/cache triggers it.

Extracted the CF Rulesets ensure/diff/PUT logic shared with
ensure-image-cdn-redirect.mjs into scripts/lib/cf-redirect-rules.mjs
to avoid duplicating it a second time (per repo sibling-pattern rule).

Rule applied live and verified: cdn.../fonts/space-grotesk-latin.woff2
now 301s to the apex, which returns 200 font/woff2.
